+++ src/CirculationLogList/CirculationLogList.test.js
@@ -1,5 +1,9 @@
 import React from 'react';
-import { render, screen, within } from '@testing-library/react';
+import {
+  render,
+  screen,
+  within,
+} from '@folio/jest-config-stripes/testing-library/react';
 
 import { CirculationLogList } from './CirculationLogList';